backend/requirements.txt
#Core framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

#Data validation & settings
pydantic[email]==2.5.0
pydantic-settings==2.1.0

#Database & ORM
sqlalchemy==2.0.23
sqlmodel==0.0.16
psycopg2-binary==2.9.9   # PostgreSQL driver
alembic==1.13.0          # Migrations

#Authentication & security  
python-multipart==0.0.6  # Form data parsing
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4   # Password hashing
bcrypt==4.0.1            # Pin to avoid passlib incompatibility with bcrypt>=4.1

#Environment & utilities
python-dotenv==1.0.0
pillow==10.1.0           # Image handling
numpy>=1.26.0           # Optional, for data ops

#Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1            # Async HTTP client
black==23.12.0           # Code formatter
ruff==0.1.8              # Linter